[Refactor] 주석 삭제 및 상수화, 변수변경
list.values -> entries로 변경,함수명 변경, 조건문 수정
