+++ src/main/kotlin/christmas/SaleController.kt
@@ -3,22 +3,21 @@ package christmas
 import christmas.data.ChristmasDiscountDayInfo
 import christmas.data.Order
 import christmas.domain.MenuSaleCalculator
-import christmas.domain.SumSaleCalculator
+import christmas.domain.TotalPriceDiscountCalculator
 
 class SaleController(private val inputDay: Int, private val menuList: List<Order>) {
 
     //입력한 요일을 판단하는 함수 , 검증 된 날짜와 검증된 Order(name,count)형식의 메뉴리스트를 받는다.
     fun sumSaleMoney() {
         val menuDiscountDay = checkDayForMenuDiscount()
-        val sumDiscountDay = checkDayForSumDiscount()
+        val sumDiscountDay = checkDayForSpecialDiscount()
         MenuSaleCalculator(menuDiscountDay).calculateDiscountRate(menuList)
-        SumSaleCalculator(sumDiscountDay).calculateDiscountSpecialDay()
+        TotalPriceDiscountCalculator(sumDiscountDay).calculateDiscountSpecialDay()
     }
 
-    private fun checkDayForSumDiscount(): ChristmasDiscountDayInfo {
+    private fun checkDayForSpecialDiscount(): ChristmasDiscountDayInfo {
         return when {
             inputDay % 7 == 3 -> ChristmasDiscountDayInfo(STAR_DAY, inputDay)
-            inputDay < 25 -> ChristmasDiscountDayInfo(CHRISTMAS_EVE_DAY, inputDay)
             inputDay == 25 -> ChristmasDiscountDayInfo(CHRISTMAS_DAY, 25)
             else -> ChristmasDiscountDayInfo(NONE, 0)
         }
@@ -35,7 +34,6 @@ class SaleController(private val inputDay: Int, private val menuList: List<Order
     companion object {
         const val STAR_DAY = "STAR_DAY"
         const val CHRISTMAS_DAY = "CHRISTMAS_DAY"
-        const val CHRISTMAS_EVE_DAY = "CHRISTMAS_EVE_DAY"
         const val NONE = "NONE"
     }
 }
