+++ src/main/kotlin/christmas/ValidateEventConditions.kt
@@ -15,7 +15,7 @@ class ValidateEventConditions(private val selectedMenus: List<Order>) {
         //음료만 주문한 경우 함수명에 맞게 true반환
         var menuCount = 0
         for (order in selectedMenus) {
-            val menu = Beverage.values().find { it.drinkName == order.menuName }
+            val menu = Beverage.entries.find { it.drinkName == order.menuName }
 
             if (menu != null) {
                 // 주문한 메뉴가 에피타이저Menu에 있는 경우
